+++ saleor/channel/models.py
@@ -1,3 +1,5 @@
+from datetime import timedelta
+
 from django.conf import settings
 from django.db import models
 from django_countries.fields import CountryField
@@ -35,6 +37,10 @@ class Channel(models.Model):
         null=True,
     )
     expire_orders_after = models.IntegerField(default=None, null=True, blank=True)
+
+    delete_expired_orders_after = models.DurationField(
+        default=timedelta(days=60),
+    )
 
     class Meta:
         ordering = ("slug",)