+++ saleor/plugins/avatax/tests/test_tasks.py
@@ -1,3 +1,4 @@
+import os
 from dataclasses import asdict
 from urllib.parse import urljoin
 
@@ -22,8 +23,8 @@ def test_api_post_request_task_sends_request(
     site_settings.company_address = address_usa
     site_settings.save()
     config = AvataxConfiguration(
-        username_or_account="",
-        password_or_license="",
+        username_or_account=os.environ.get("AVALARA_USERNAME", ""),
+        password_or_license=os.environ.get("AVALARA_PASSWORD", ""),
         use_sandbox=True,
         from_street_address="Tęczowa 7",
         from_city="WROCŁAW",
@@ -54,8 +55,8 @@ def test_api_post_request_task_creates_order_event(
     site_settings.save()
 
     config = AvataxConfiguration(
-        username_or_account="",
-        password_or_license="",
+        username_or_account=os.environ.get("AVALARA_USERNAME", ""),
+        password_or_license=os.environ.get("AVALARA_PASSWORD", ""),
         use_sandbox=True,
         from_street_address="Tęczowa 7",
         from_city="WROCŁAW",